feat(send/contactlist): show typing state in message preview
Fold real-time typing indicators into the send-flow contact list, and pay
down the accompanying bloat in the list model and ViewModel:

- Regroup the chat-derived fields on ContactListItem.ContactRow into a
  nullable Conversation (chatId, lastMessagePreview, unreadCount, isTyping).
  lastActivity stays on the row since it's also the sort key for on-Flipcash
  contacts that have no chat yet.
- Extract generateListItems/formatPreview out of SendFlowViewModel into a
  pure, injectable ContactListBuilder, shrinking the ViewModel and making the
  list logic unit-testable on its own.
- Apply typing as a cheap second-stage flow overlay on top of the base list
  (distinctUntilChanged typing keys, self filtered out) so typing ticks never
  re-filter or re-sort the list.
